Update PHPicker.swift
Switched out a single `Image?` as the selected image with an array of `Data` objects. This allows for selection of multiple images (set via configuration) uses the new algorithm for loading the selection images (db15627).

Binding the selected images as `Data` objects also allows for more flexible use on the other end. The developer might just want the `Data` rather than `Image`s (which can't be turned back into `Data`) or `UIImage`s (which can be super bloated).