Update PHPicker.swift
Now that it's known that selected Live Photos return `true` when asked if they can be loaded as a standard image, things have been refactored to check if it's a Live Photo first. Plus, there's a new local property that overrides if Live Photos should just be flattened into regular photos (`keepLivePhotosIntact`).