Update PHPicker.swift
Refactored `PHPicker` to no longer use a `Binding`. This could have lead to confusion as to how to know when it's done. A modifier to could be used to listen to changes to `selections`, but then it might get called again when the bound property is reset back to empty. Just in case, a completion handler will be clearer. And using the `Result` type, it can convey if there was an underlying issue.